backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
websockets
python-dotenv
python-jose
google-auth
requests
pydantic
pypdf
python-docx
openai-agents
